requirements.txt
streamlit==1.41.1
streamlit-drawable-canvas==0.9.3
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.3
Pillow==10.4.0
